+++ domain/src/main/java/tachiyomi/domain/library/interactor/SearchLibrary.kt
@@ -1,0 +1,16 @@
+package tachiyomi.domain.library.interactor
+
+import tachiyomi.domain.library.model.search.LibrarySearchLexer
+import tachiyomi.domain.library.model.search.LibrarySearchParser
+import tachiyomi.domain.library.repository.LibrarySearchRepository
+
+class SearchLibrary(
+    private val librarySearchRepository: LibrarySearchRepository,
+) {
+    suspend fun await(searchQuery: String): Set<Long> {
+        val tokens = LibrarySearchLexer.tokenize(searchQuery)
+        val rootNode = LibrarySearchParser(tokens).parse()
+
+        return librarySearchRepository.getFilteredMangaIdsByAst(rootNode)
+    }
+}
